Add Go as a new language supported by Hound.
This adds basic support to Hound for Go,
by integrating against [hound-go]
via Hound's external review Resque API.

This should be considered an early release feature.
As the review service matures
and the feature-set stabilizes,
we can define supported configuration options
and the necessary documentation
to promote this to an officially supported language.

  [hound-go]: github.com/thoughtbot/hound-go